+++ Tests/New-CT365Group.Tests.ps1
@@ -1,11 +1,14 @@
 BeforeAll {
-    . $PSScriptRoot/New-CT365Group.ps1
+    # Call Cmdlet
+    $commandScriptPath = Join-Path -Path $PSScriptRoot -ChildPath '..\functions\public\New-CT365Group.ps1'
+
+    . $commandScriptPath
 }
 
 Describe 'New-CT365Group Function' {
     Context 'When provided invalid parameters' {
         It 'Should throw an error for invalid domain format' {
-            $filePath = "C:\Path\to\file.xlsx"
+            $filePath = $commandScriptPath
             $domain = "invalid_domain"
             
             { New-CT365Group -FilePath $filePath -Domain $domain } | Should -Throw