+++ CHANGELOG.md
@@ -1,5 +1,12 @@
-### 365AutomatedLab Changelog
+# 365AutomatedLab Changelog
 
+## 0.1.3
+
+Fixed issue with New-CT365SharePointSite not creating each of the different sites correctly every time
+
+## 0.1.2
+
+Updated/Created Pester Tests
 
 ## 0.1.1